# resources/js/api/models/showUserResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * tasker
 * OpenAPI spec version: 0.1
 */
import type { User } from "./user";

/**
 * Show user response object.
 */
export interface ShowUserResponse {
  data: User;
  message: string;
}