# resources/js/api/models/user.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * tasker
 * OpenAPI spec version: 0.1
 */

/**
 * User resource object
 */
export interface User {
  id: number;
  name: string;
  /** @nullable */
  email: string | null;
}